`update_video_num`移到`view_utils`
